the-stick-of-condemnation: Events and animations
Move styles to styles.css, dialogue and event handling in script.js to
update css styles on animationEnd event of the #message element.

The animationEnd event allows for continued looping of a css animation
to trigger side effects.

Each time the "typing" animation ends, change styles related to who is
talking, updated the html content of the dialogue box, and reapply the
"typing" class to make the animation start again.