the-stick-of-condemnation: Vary animation length
Animations can be variable lengths. They are assigned a class-{n}s
where in is the length of time to play the animation.

A function getReadingTime(str) returns an estimated reading time
for a sentence. The delay: property on messages can be used
to add dramatic pauses to the script by adding the delay to the
estimated reading time.